refactor(sandbox): remove tool policy facade
